Added second hang-detection algorithm
This one is a bit more advanced. It also tries to filter out changes that are not effective. This includes:
- Changes to data values away from zero
- Shifts between a limited set of data values

As soon as it detects any of the following, it is not considered a hang:
- A value becoming zero (this implies different branching)
- A new data value being entered (in the sample period) that is not-zero. The not-zero criterion is added to detects of programs generating run-away sequences.

It may still require some tweaking, as it appears to consider the #13549 7x7 Busy Beaver a hang now.