Tweaked Hang Detection 2 algorithm
It is now less aggressive, as it also considers data value changes from zero as significant. This is much safer/easier, but prevents hang-detection for run-away sequence generators.

For this, however, a smaller data size could be useful, which has therefore been reduced. It now finds the #13549 7x7 Busy Beaver again.