Fix to Hang Detection 2
The check to see if new data cells were visited compared to the last time was incorrect. The bounds were set relative to the data pointer at the start which meant they could not be directly compared. This has been fixed.

Also added more logging to analyze why it incorrectly classified certain long-running programs as hangs. The cause has been found but is yet to be fixed.