Optimized Program access and Program Pointer manipulation
Changed Program Pointer to directly point into the program storage.
Changed directory values so that they can be directly added to the Program Pointer.
To enable this, program storage is now based on maximum size instead of actual size.
Furthermore, extended storage so that it also allows access when the Program Pointer is just outside of the program to quickly detect program determination.