Only invoke resetHangDetection at start of run
It is now up to the hang detectors to reset their status after isHangDetected is invoked.
For the execution hang detection the reset was relatively expensive, so it has been changed to clear the previous array while it is being iterated for comparison. The drawback is that hangs are potentially detected later.