Improve Data hang detection to exclude some false positives
Extended detection to also take into account the overall direction of the program pointer in the sample period.
This is required to avoid false positives for programs that go back and forth over long stretches of the data tape, where traversal in one direction is relatively slow and does not modify any data.

Then the following can happen:
- in sample period A the data pointer moved in the fast direction, possibly modifying some data cells along the way and turned, moving a bit in the slow direction
- the next sample period, B, the data pointer keeps moving in the slow direction, not modifying any data cells, nor reaching the end.

In such a case, the following can happen:
- No new program instructions are executed
- No data cells are modified
- No new new data cells are entered

This, however, should not be signalled as a hang. This can be achieved by also considering the overall direction that the data pointer moved.