Remove "max steps per run" setting
There is no good reason to have a "max steps per run" setting next to the more intuitive "max total steps" setting.
It was mainly a side-effect of the implementation to check for termination periodically instead of every step (as a minor optimization).
This optimization is now hidden by updating the data stack size to reflect that it can be more than the maximum number of steps.

Introduced SearchSettings struct.
This ensures that re-allocation after changing search settings only needs to be done once, which is useful now that the data stack size depends on two search settings.