Do not terminate periodic hang check prematurely
When a periodic hang check is initiated (with an assumed periodic cycle) only abort it when the hang check failed.

So do not let it abort simply by considering the number of steps carried out since initiating the check. This prevented detection of hangs with a period close to the hang sample period (as three periods need to be executed before some types of hang can be detected).