Made cycle detection usage more generic
It is now enabled via dedicated flag (to reduce coupling with specific hang detection method).
Also replaced _opsToWaitBeforePeriodicHangCheck by _periodicHangCheckAt to more cleanly separate generic search logic from hang detection. The newly added getNumRecordedInstructions method is now used to determine when to check for a periodic hang.
This changed the timing of the checks a bit, which caused some hangs to not be detected anymore.
Added new 6x6-DelayedHang test for this. Bumping the number of periodic check attempts fixed this test. This in turn delayed the sweep tests, requiring an extension of the maximum number of steps in certain tests.
This triggered further tuning of 6x6 OrchestratedSearch test, reducing the number of undetected hangs from 3325 to 2335.