Let ExhaustiveSearcher directly signal left turns to the hang detector
This facilitates sweep hang detection.
It avoids having to handle the following cases when hang detection is invoked after the execution step has been completed:
- The value that is the turn is not zero anymore (due to a INC/DEC operation)
- The data pointer does not point to the boundary value anymore (due to a SHR/SHL operation)

Also added checks to fail the hang detection when the sweep shrinks.
This should ensure that the current check does not detect "glider" hangs.
Although these are hangs as well, this check should not identify those (as this could also lead to false positives).