Added DeltaTracker
The DeltaTracker is initially used by the RegularSweepHangDetector to see if any values are skipped without being evaluated, replacing the custom logic.
As a next step, it will be used by the detector to adapts its detection accordingly, instead of failing it.

As a result of this change, slightly fewer hangs are detected by the 6x6 search. Why that is still needs to be investigated.