Fixed No Exit Hang Detection
When checking if a path already had been followed, it did not take into account the data value, which meant it did not follow all paths and wrongly concluded that the loop could not be exited.
This is fixed now.

With this an additional 1109 hangs are now detected. Updated 6x6 test accordingly.
Also added tests for some of these newly detected No Exit hangs.