Added ExitFinder
It is similar to, and will replace, the NoExitHangDetector.
The difference is that it uses the compiled program to quickly detect no exit hangs.

The NoExitHangDetector is not yet removed.
There are still some differences in results that need to be investigated.

Also fixed a problem in CompiledProgram.
Popping the state did not always result in a correct state, as finalization of blocks was not undone.
This could lead to blocks referencing blocks that did not exist anymore.