Removed old NoExitHangDetector
Extended ExitFinder so that it finds all the hangs detected by the old detector, plus some more.

This required extending the implementation so that it finalizes blocks that can be finalized given the instructions that have been set (but where this has not yet been done as PP has not yet entered them and never may).

Furthermore, the CompiledProgram class has been made smarter.
It now detects when it is not possible to exit a program block via the zero exit, and will then not instantiate this block.

Updated No Exit Hang Tests:
- Added a few tests for hangs that were previously not detected, but are now detected by the ExitFinder
- Removed 6x6-InnerSquare test, as from the Exit analysis it cannot actually be concluded that the program hangs.

Reduced total undetected hangs for 6x6 search test to 1564.