Introduced new MetaPeriodicHangDetector for meta-level periodic hangs
Split the original implementation in a base class (for hangs at the lowest-level run summary) and a subclass (for hangs at the meta-level run summary).
This was done so that assumed hangs at the meta-level, which actually switched to a hang at the lowest level can be detected.
That this situation could occur was already identified in the comment, but an example 6x6 program that exhibited this behaviour was found.
Added new test, 6x6-PeriodicHangBreakingOutOfAssumedMetaLevelLoop, which now passes.