Detect last 6x6 hang that still caused an error
Extended Sweep Hang detector to check adherence to the sweep contract.
This way, it will (eventually) fail when the hang is of another type so that it can be detected by another detector.

Added new 6x6-PeriodicHangBreakingOutOfAssumedMetaLevelLoop3 test, which is now correctly detected as a periodic hang.

Updated expected results of 6x6 search test.