Extended Sweep Hang detector to ignore fixed loops during reversal
When the Sweep Hang detector starts, it now checks in more detail if the current loop  matches that of a Sweep Hang.
In particular, it should contain (a multiple of) two sweep loops.
Loops with a fixed number of iterations (which can be part of the reversal sequence) are now recognized and ignored.

With this, the hang in 6x6-RegularSweepWithReversalContainingFixedLoop is now recognized by the Sweep Hang detector.

It only bumps the total number of detected 6x6 hangs by three.
Nevertheless, the 6x6 search test is updated accordingly.