Added condensed dump for RunSummary
In doing so fixed the following: Loop period was not stored correctly for run blocks

Also changed how program blocks are stored in core run summary.
Now storing program block index instead of start index (on program grid).
This maps more easily to the compiled program and is therefore easier for dissemination.