First attempt to improve No Exit detection algorithm
When a program block is found that cannot yet be finalized (which is a possible exit out of its current execution) it is now checked if this exit can be reached.

It does not yet work as intended.
It currently signals a few false positives.
It seems the exit detection corrupts the interpreted program.
This requires further investigation.