Fixed issue with No Exit hang false positives
The problem was not with the newly added reachability analysis, but due to a missing enterBlock invocation.

Added new test of program that was wrongly reported as a hang due to this bug.
Moved some test from failing hang tests to No Exit hang tests.
Updated total of undetected hangs of 6x6 search. It is now down to 571.