Added Fast Execution
Added new "max steps" search parameter.
The original parameter is used to determine when abort program execution (and assume a hang).
The new parameter is used to determine when to disable hang detection.
When the number of execution steps has exceeded this limit, or when the run summary buffer is full, hang detection is disabled.
Execution then switches to use the new FastExecutor.
This runs more quickly as it does not need to track any state required for hang detection, nor any undo history.
A drawback is that once in this mode, programs cannot be further expanded by the search.
It can therefore identify promising instruction stacks (programs that break out of their known execution) but does not generate (all) possible programs that could result from this.
These "late escapes" are however reported, and can be investigated later.