Start implementation of StaticMetaPeriodicHangDetector
Changed invocation of static hang detection. It only requires a loop iteration to have finished. Whether or not it starts a new iteration is passed to the detector. RunSummary has been updated to more clearly distinguish loop end and loop start.

Extended StaticHangDetector to allow for initial indecisiveness when checking the hang behavior.

Initial implementation of StaticMetaPeriodicHangDetector includes implementation of exhibitsHangBehaviour method. Implementation of analyseLoop is still pending.