Integrated new StaticMetaPeriodicHangDetector
Changed when static hang detection is invoked by ExhaustiveSearcher as both detectors have different requirements. Detection is now always invoked after an instruction that is part of a loop was executed.

Changed API of StaticHangDetector. Initializing loops should now be done when the behavior is analyzed. The key methods are now:
- bool shouldCheckNow(bool loopContinues)
- bool analyzeHangBehaviour()
- Trilian proofHang()

Also simplified StaticHangDetector implementation. Removed use of stage member variable.

Changed implementation of StaticMetaPeriodicHangDetector. Instead of verifying that last two meta-loop iterations are equal, it goes as far back as possible. This is required to ensure the meta-loop is fully bootstrapped.

All tests pass again.