Improve loop analysis for travelling loops
Calculate the number of bootstrap cycles more accurately.
It could be less than what was originally calculated.
It should be based on the instruction that will mask the exit condition of the
bootstrap-only exit.

Also express the exit-condition wrt to the value as it is first consumed by the loop.
This will help in analysing exit conditions of bootstrap loops.
It also makes it more clear why certain exits can only occur during bootstrap.