Improve accuracy of periodic hang detector
For travelling loops, the detector could falsely signal a hang.
After the loop was bootstrapped, it only checked that all data values yet to be consumed were zero.
However, it could theoretically happen that some already consumed values still cause the loop to exit.
This case is now covered by letting the loop run for longer before concluding it is hanging.

This unfortunately requires tracking extra state, which adds more complexity than desired.

On the other hand, the efficiency of the meta-periodic check is improved.
It used to analyze the same loop repeatedly.
Not anymore.