Let sweep hang detector only detect regular hang patterns
Restored condition that program should be in a meta-loop.
This avoids problems with determining how far to analyze the sequence.
Too far can cause hang detection to fail (as it includes execution before the hang).
Not far enough may cause false positives.

Also moved tests for irregular hangs to a dedicated test file.