Extend support for rotation-equivalent sweep loops
When there can be more than one variant of a loop, it matters which one preceded a given transition, in order to correctly establish the exit instruction.
This is now done.

For simplicity (and extensibility), each variant is analysed separately.
This means that a transition group can now contain more than one loop analysis.