Fix the final failing Sweep Hang Test
Let zeroes ahead check ignore value as specified DP.

Also remove check for non-exit values for the loop in opposite direction.
It was far too basic.
It should also include changes made by the transition in a future iterations, as well as changes made by the sweep loop, including bootstrap-only changes.
It can be re-introduced when there are actual programs where this check is needed to avoid a false positive.