Improve support for inside-sweep changes made by transitions
Also allow changes where the transition delta goes against the delta of the sweep sequence change, as long as it only dampens it but does not change the direction.
This detects another ten 6x6 sweep hangs.

Added a few more tests for regular sweep hangs that are not yet detected.
Also improved assert for regular sweep hang test that the hang type is as expected.