Improve sweep end-type classification
In analyzing irregular sweeps, discovered that one hang was actually a regular one, but late to enter this meta-loop.
It results in an irregular growth.
Improved sweep end-type classification accordingly.

Also generalized related condition. It checks that all transitions are on zero, instead of requiring this number fo be one.
This results in detection of another interesting hang.

So in total two more 6x6 sweep hangs are detected.