Fix assertion failure during 7x7 search
Search found a program that was executing a meta-loop, with sweep behavior with transitions still differing between iterations.
This program has been added as a new test (whose hang is not yet detected).
Removed assertion.

During analysis discovered that there can be sweeps where there is a mid-sweep transition but where the outgoing loop and incoming loop are the same.
This was not yet supported.
Analysis has been updated to support this.