Extend meta-loop analysis with data pointer deltas
Also change loop iteration delta.
It is now with respect to the execution of the run block in the
previous iteration of the meta-loop.
This makes a difference for meta-loop analysis where more than one
loop period is analyzed, e.g. TwoStateRightSweep-SharedTransition

Also fix pooling of sequence and loop analysis objects.
The raw pointer approach failed when the background storage used by
the pooling vectors was resized.
Also, the clearing of the pools meant that the dynamic memory they
relied on was not re-used.