Fix LoopBehavior::loopType
Loops without a min DP and without a max DP were wrongly classified
as gliders.
This in turn caused the assertion failure in
LoopAnalysis::stationaryLoopExits for
6x6-IrregularSweepWithAlternatingEndSweepTransition.

With this fix, several irregular sweep hangs (including the above)
are not detected anymore.
So increase total of undetected hangs to 12.

Also fixed various minor errors in comments.